+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Dodano obsługę Money w Range
 - Merge z branchem 1.0 (do tagu vX.Y.Z)
 - Dodano opcje `min_options` i `max_options` w `RangeType`
+- Dodano `PercentRangeValidator`
 
 #### v1.1.3 (2013-07-09) ####
 - Merge z branchem 1.0 (do tagu v1.0.4)